pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hedge-fund-ai"
version = "3.0.0"
description = "Production-ready multi-agent AI financial analysis — hedge-fund grade"
requires-python = ">=3.11"

dependencies = [
    # ── Web framework ─────────────────────────────────────────────────────────
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",

    # ── Config & validation ───────────────────────────────────────────────────
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",

    # ── LangGraph + LangChain ────────────────────────────────────────────────
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",

    # ── LLM structured output ─────────────────────────────────────────────────
    "openai>=1.35.0",
    "instructor>=1.3.0",
    "google-generativeai>=0.8.0",
    "tenacity>=8.3.0",

    # ── RAG / vector store ────────────────────────────────────────────────────
    "langchain-chroma>=0.1.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "tiktoken>=0.7.0",

    # ── Technical analysis ────────────────────────────────────────────────────
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "ta>=0.11.0",

    # ── Semantic cache ────────────────────────────────────────────────────────
    "redisvl>=0.3.0",
    "redis>=5.0.0",

    # ── Database ──────────────────────────────────────────────────────────────
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "aiosqlite>=0.20.0",

    # ── Observability ─────────────────────────────────────────────────────────
    "arize-phoenix[otel]>=4.0.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp-proto-http>=1.24.0",
    "openinference-instrumentation-openai>=0.1.12",

    # ── Rate limiting ─────────────────────────────────────────────────────────
    "slowapi>=0.1.9",

    # ── Evaluation ────────────────────────────────────────────────────────────
    "deepeval>=1.0.0",

    # ── Utilities ─────────────────────────────────────────────────────────────
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
